LMS: make array indexes size_t
Avoid unsigned int indexes.

+++ WsmanClient/AuditLogWSManClient.h
@@ -1,6 +1,6 @@
 /* SPDX-License-Identifier: Apache-2.0 */
 /*
- * Copyright (C) 2009-2023 Intel Corporation
+ * Copyright (C) 2009-2024 Intel Corporation
  */
 /*++
 
@@ -69,16 +69,16 @@ private:
 	std::string DisplayExtendedData(unsigned short appId, unsigned short eventId, std::vector<uint8_t> extendedData,
 		uint8_t extendedDataLen);
 	std::string PrintOptInPolicy(uint8_t curData, const char* title);
-	std::string PrintUint32(uint8_t* extData, uint8_t extendedDataLen, const char* message, int & i);
-	std::string PrintUint16(uint8_t* extData, uint8_t extendedDataLen, const char* message, int & i);
-	std::string PrintInterfaceHandleUint32(uint8_t* extData, uint8_t extendedDataLen, int & i);
-	std::string DisplayWirelessProfileName(uint8_t* extData, uint8_t extendedDataLen, int & i);
+	std::string PrintUint32(uint8_t* extData, uint8_t extendedDataLen, const char* message, size_t &i);
+	std::string PrintUint16(uint8_t* extData, uint8_t extendedDataLen, const char* message, size_t &i);
+	std::string PrintInterfaceHandleUint32(uint8_t* extData, uint8_t extendedDataLen, size_t &i);
+	std::string DisplayWirelessProfileName(uint8_t* extData, uint8_t extendedDataLen, size_t &i);
 	std::string DisplayBasicUsernameSidInformation(uint8_t* extData, uint8_t extendedDataLen, unsigned short numOfTabsToIdent, const char *action);
 	std::string DisplayAuthenticationMode (int status);
 	std::string DisplayAuthenticationStatus (int status);
 	std::string DisplayFirmwareVerion(uint8_t* extData, int i);
 	std::string DisplayTimeStamp(uint8_t* extData);
-	std::string DisplayFullWirelessProfileName(uint8_t* extData, uint8_t extendedDataLen, int & i);
+	std::string DisplayFullWirelessProfileName(uint8_t* extData, uint8_t extendedDataLen, size_t &i);
 	std::string DisplayAgentID(uint8_t* extData);
 	std::string DisplaySecurityAdminAmtProvisioningCompletedEvent(uint8_t* extData, uint8_t extendedDataLen);
 	std::string DisplaySecurityAdminAclEntryAddedEvent(uint8_t* extData, uint8_t extendedDataLen);
